+++ src/lib/output/themes/default/partials/index.tsx
@@ -1,7 +1,7 @@
 import { classNames, renderName } from "../../lib";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext";
-import { JSX, Raw } from "../../../../utils";
-import { ContainerReflection, DeclarationReflection, ReflectionCategory, ReflectionKind } from "../../../../models";
+import { JSX } from "../../../../utils";
+import type { ContainerReflection, ReflectionCategory } from "../../../../models";
 
 function renderCategory(
     { urlTo, icons, getReflectionClasses }: DefaultThemeRenderContext,
@@ -72,15 +72,6 @@ export function index(context: DefaultThemeRenderContext, props: ContainerReflec
 
     return (
         <>
-            {props instanceof DeclarationReflection &&
-                props.kind === ReflectionKind.Module &&
-                props.readme?.length !== 0 && (
-                    <section class="tsd-panel-group">
-                        <section class="tsd-panel tsd-typography">
-                            <Raw html={context.markdown(props.readme || [])} />
-                        </section>
-                    </section>
-                )}
             <section class="tsd-panel-group tsd-index-group">
                 <section class="tsd-panel tsd-index-panel">{content}</section>
             </section>